+++ commands/Kernel.d.ts
@@ -1,5 +1,4 @@
 import type { Command } from "commander";
 export default class Kernel {
     static registerCommands(program: Command): void;
-    private static commands;
 }