+++ middlewares/MaintenanceMiddleware.d.ts
@@ -1,0 +1,4 @@
+import type { HandlerType } from "../types/router";
+export default class MaintenanceMiddleware {
+    handle(handler: HandlerType): HandlerType;
+}