+++ src/builders/RouterBuilder.ts
@@ -153,7 +153,7 @@ export default class RouterBuilder {
             throw new RouterInvalidException(`Invalid router controller definition: ${definition}.`);
         }
 
-        const controllerPath = path.resolve(App.rootPath(), this.baseNamespace);
+        const controllerPath = path.resolve(App.Path.rootPath(), this.baseNamespace);
         const location = Bun.resolveSync(`./${controllerName}.ts`, controllerPath);
 
         let ControllerClass: any;