+++ src/commands/maintenance/MaintenanceDownCommand.ts
@@ -1,4 +1,4 @@
-import App from "@bejibun/app";
+import AppConfig from "@bejibun/app/config/app";
 import Logger from "@bejibun/logger";
 import {DateTime} from "luxon";
 
@@ -34,7 +34,7 @@ export default class MaintenanceDownCommand {
     protected $arguments: Array<Array<string>> = [];
 
     public async handle(options: any, args: Array<string>): Promise<void> {
-        await Bun.write(App.storagePath("framework/maintenance.down.json"), JSON.stringify({
+        await Bun.write(AppConfig.maintenance.file, JSON.stringify({
             message: "🚧 We're doing maintenance. Please check back soon.",
             status: 503,
             allows: options.allows,