+++ src/commands/maintenance/MaintenanceUpCommand.ts
@@ -1,0 +1,38 @@
+import AppConfig from "@bejibun/app/config/app";
+import Logger from "@bejibun/logger";
+
+export default class MaintenanceUpCommand {
+    /**
+     * The name and signature of the console command.
+     *
+     * @var $signature string
+     */
+    protected $signature: string = "maintenance:up";
+
+    /**
+     * The console command description.
+     *
+     * @var $description string
+     */
+    protected $description: string = "Turn app into live mode";
+
+    /**
+     * The options or optional flag of the console command.
+     *
+     * @var $options Array<Array<any>>
+     */
+    protected $options: Array<Array<any>> = [];
+
+    /**
+     * The arguments of the console command.
+     *
+     * @var $arguments Array<Array<string>>
+     */
+    protected $arguments: Array<Array<string>> = [];
+
+    public async handle(options: any, args: Array<string>): Promise<void> {
+        if (await Bun.file(AppConfig.maintenance.file).exists()) await Bun.file(AppConfig.maintenance.file).delete();
+
+        Logger.setContext("APP").info("Application turned into live mode.");
+    }
+}