+++ src/server.ts
@@ -2,9 +2,11 @@ import App from "@bejibun/app";
 import Logger from "@bejibun/logger";
 import RuntimeException from "@/exceptions/RuntimeException";
 import Router from "@/facades/Router";
-import "@/bootstrap";
+import {RouterGroup} from "@/types";
+import MaintenanceMiddleware from "@/middlewares/MaintenanceMiddleware";
+import (App.Path.rootPath("bootstrap.ts"));
 
-const exceptionHandlerPath = App.appPath("exceptions/handler.ts");
+const exceptionHandlerPath = App.Path.appPath("exceptions/handler.ts");
 let ExceptionHandler: any;
 try {
     ExceptionHandler = require(exceptionHandlerPath).default;
@@ -12,16 +14,16 @@ try {
     throw new RuntimeException(`Missing exception handler class [${exceptionHandlerPath}].`);
 }
 
-const apiRoutesPath = App.routesPath("api.ts");
-let ApiRoutes: any;
+const apiRoutesPath = App.Path.routesPath("api.ts");
+let ApiRoutes: RouterGroup;
 try {
     ApiRoutes = require(apiRoutesPath).default;
 } catch {
     throw new RuntimeException(`Missing api file on routes directory [${apiRoutesPath}].`);
 }
 
-const webRoutesPath = App.routesPath("web.ts");
-let WebRoutes: any;
+const webRoutesPath = App.Path.routesPath("web.ts");
+let WebRoutes: RouterGroup;
 try {
     WebRoutes = require(webRoutesPath).default;
 } catch {
@@ -42,13 +44,15 @@ const server = Bun.serve({
     port: Bun.env.APP_PORT,
 
     routes: {
-        ...Router.namespace("app/exceptions").any("/*", "Handler@route"),
+        "/": require(App.Path.publicPath("index.html")),
 
-        "/": require(App.publicPath("index.html")),
+        ...Router.middleware(new MaintenanceMiddleware()).group([
+            Router.namespace("app/exceptions").any("/*", "Handler@route"),
 
-        ...ApiRoutes,
+            ApiRoutes,
 
-        ...WebRoutes
+            WebRoutes
+        ])
     }
 });